Update allowed hosts url- settings.py
Import add_item function from views.py and add path named 'add' as well as add_item function. Name it 'add'- urls.py

Import Item model from models.py and register the model on the admin site-admin.py

Create forms.py file and import forms from Django framework, as well as Item model. Create ItemForm class that uses the buily in Django forms functionality and ModelForm. It uses the Item model and creates the fields which are name and done in order to display them to the user.

Create Item model which defines the name and done fields of the form such as max length, blank =false etc. It also displays the name of the field on the admin site at the bottom of the class- models.py

Create add_items function where it returns the form, saves it and then redirects the user to the homepage with newly added item- views.py

Add csrf token and form Jinja syntax on add_item.html page which is the form displayed to the user

Iterate through the items and display them as marked if the done checkbox has been checked, and display them as normal if not. Display 'You have nothing to do' if there are no items.- todo_list.html